+++ conn.go
@@ -6,130 +6,6 @@ import (
 	"sync"
 )
 
-type DTLSClientConnection struct {
-	UDPClientConnection
-	dtlsClient *DTLSClient
-}
-
-func (c *DTLSClientConnection) Send(req Request) (resp Response, err error) {
-	msg := req.GetMessage()
-	opt := msg.GetOption(OptionBlock1)
-
-	if opt == nil { // Block1 was not set
-		if MessageSizeAllowed(req) != true {
-			return nil, ErrMessageSizeTooLongBlockOptionValNotSet
-		}
-	} else { // Block1 was set
-		// log.Println("Block 1 was set")
-	}
-
-	if opt != nil {
-		blockOpt := Block1OptionFromOption(opt)
-		if blockOpt.Value == nil {
-			if MessageSizeAllowed(req) != true {
-				err = ErrMessageSizeTooLongBlockOptionValNotSet
-				return
-			} else {
-				// - Block # = one and only block (sz = unspecified), whereas 0 = 16bits
-				// - MOre bit = 0
-			}
-		} else {
-			payload := msg.GetPayload().GetBytes()
-			payloadLen := uint32(len(payload))
-			blockSize := blockOpt.BlockSizeLength()
-			currSeq := uint32(0)
-			totalBlocks := uint32(payloadLen / blockSize)
-			completed := false
-
-			var wg sync.WaitGroup
-			wg.Add(1)
-
-			for completed == false {
-				if currSeq <= totalBlocks {
-
-					var blockPayloadStart uint32
-					var blockPayloadEnd uint32
-					var blockPayload []byte
-
-					blockPayloadStart = currSeq*uint32(blockSize) + (currSeq * 1)
-
-					more := true
-					if currSeq == totalBlocks {
-						more = false
-						blockPayloadEnd = payloadLen
-					} else {
-						blockPayloadEnd = blockPayloadStart + uint32(blockSize)
-					}
-
-					blockPayload = payload[blockPayloadStart:blockPayloadEnd]
-
-					blockOpt = NewBlock1Option(blockOpt.Size(), more, currSeq)
-					msg.ReplaceOptions(blockOpt.Code, []Option{blockOpt})
-					modifiedMsg := msg.(*CoapMessage)
-					modifiedMsg.SetMessageId(GenerateMessageID())
-					modifiedMsg.SetPayload(NewBytesPayload(blockPayload))
-
-					// send message
-					_, err2 := c.sendMessage(modifiedMsg)
-					if err2 != nil {
-						wg.Done()
-						return
-					}
-					currSeq = currSeq + 1
-
-				} else {
-					completed = true
-					wg.Done()
-				}
-			}
-		}
-	}
-
-	resp, err = c.sendMessage(msg)
-	return
-}
-
-func (c *DTLSClientConnection) sendMessage(msg Message) (resp Response, err error) {
-	if msg == nil {
-		return nil, ErrNilMessage
-	}
-
-	b, err := MessageToBytes(msg)
-	if err != nil {
-		return
-	}
-
-	_, err = c.dtlsClient.Write(b)
-	if err != nil {
-		return
-	}
-
-	if msg.GetMessageType() == MessageNonConfirmable {
-		resp = NewResponse(NewEmptyMessage(msg.GetMessageId()), nil)
-		return
-	}
-
-	// c.conn.SetReadDeadline(time.Now().Add(2))
-
-	msgBuf := make([]byte, 1500)
-	n, err := c.dtlsClient.Read(msgBuf)
-	if err != nil {
-		return
-	}
-
-	respMsg, err := BytesToMessage(msgBuf[:n])
-	if err != nil {
-		return
-	}
-	resp = NewResponse(respMsg, nil)
-
-	return
-}
-
-func (c *DTLSClientConnection) Close() {
-	c.dtlsClient.Close()
-}
-
 func MessageSizeAllowed(req Request) bool {
 	msg := req.GetMessage()
 	b, _ := MessageToBytes(msg)
@@ -141,11 +17,11 @@ func MessageSizeAllowed(req Request) bool {
 	return true
 }
 
-type UDPClientConnection struct {
+type UDPConnection struct {
 	conn net.Conn
 }
 
-func (c *UDPClientConnection) ObserveResource(resource string) (tok string, err error) {
+func (c *UDPConnection) ObserveResource(resource string) (tok string, err error) {
 	req := NewRequest(MessageConfirmable, Get, GenerateMessageID())
 	req.SetRequestURI(resource)
 	req.GetMessage().AddOption(OptionObserve, 0)
@@ -156,7 +32,7 @@ func (c *UDPClientConnection) ObserveResource(resource string) (tok string, err 
 	return
 }
 
-func (c *UDPClientConnection) CancelObserveResource(resource string, token string) (err error) {
+func (c *UDPConnection) CancelObserveResource(resource string, token string) (err error) {
 	req := NewRequest(MessageConfirmable, Get, GenerateMessageID())
 	req.SetRequestURI(resource)
 	req.GetMessage().AddOption(OptionObserve, 1)
@@ -165,20 +41,19 @@ func (c *UDPClientConnection) CancelObserveResource(resource string, token strin
 	return
 }
 
-func (c *UDPClientConnection) StopObserve(ch chan ObserveMessage) {
+func (c *UDPConnection) StopObserve(ch chan ObserveMessage) {
 	close(ch)
 }
 
-func (c *UDPClientConnection) Close() {
-	c.conn.Close()
+func (c *UDPConnection) Close() error {
+	return c.conn.Close()
 }
 
-func (c *UDPClientConnection) Observe(ch chan ObserveMessage) {
-	conn := c.conn
+func (c *UDPConnection) Observe(ch chan ObserveMessage) {
 
 	readBuf := make([]byte, MaxPacketSize)
 	for {
-		len, err := conn.Read(readBuf)
+		len, err := c.Read(readBuf)
 		if err == nil {
 			msgBuf := make([]byte, len)
 			copy(msgBuf, readBuf)
@@ -198,7 +73,7 @@ func (c *UDPClientConnection) Observe(ch chan ObserveMessage) {
 	}
 }
 
-func (c *UDPClientConnection) Send(req Request) (resp Response, err error) {
+func (c *UDPConnection) Send(req Request) (resp Response, err error) {
 	msg := req.GetMessage()
 	opt := msg.GetOption(OptionBlock1)
 
@@ -276,7 +151,7 @@ func (c *UDPClientConnection) Send(req Request) (resp Response, err error) {
 	return
 }
 
-func (c *UDPClientConnection) sendMessage(msg Message) (resp Response, err error) {
+func (c *UDPConnection) sendMessage(msg Message) (resp Response, err error) {
 	if msg == nil {
 		return nil, ErrNilMessage
 	}
@@ -286,7 +161,7 @@ func (c *UDPClientConnection) sendMessage(msg Message) (resp Response, err error
 		return
 	}
 
-	_, err = c.conn.Write(b)
+	_, err = c.Write(b)
 	if err != nil {
 		return
 	}
@@ -299,7 +174,7 @@ func (c *UDPClientConnection) sendMessage(msg Message) (resp Response, err error
 	// c.conn.SetReadDeadline(time.Now().Add(2))
 
 	msgBuf := make([]byte, 1500)
-	n, err := c.conn.Read(msgBuf)
+	n, err := c.Read(msgBuf)
 	if err != nil {
 		return
 	}
@@ -315,6 +190,13 @@ func (c *UDPClientConnection) sendMessage(msg Message) (resp Response, err error
 
 		c.Send(NewRequestFromMessage(ack))
 	}
-
 	return
+}
+
+func (c *UDPConnection) Write(b []byte) (int, error) {
+	return c.conn.Write(b)
+}
+
+func (c *UDPConnection) Read(b []byte) (int, error) {
+	return c.conn.Read(b)
 }
