--- dtls.c
@@ -1,3 +1,0 @@
-#include "dtls.h"
-
-BIO_METHOD* go_session_bio_method;
