+++ init.go
@@ -1,5 +1,5 @@
 package canopus
 
 func init() {
-	SERVER_INSTANCES = make(map[string]CoapServer)
+
 }
